% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folds.R
\name{create_folds}
\alias{create_folds}
\title{Creates Folds}
\usage{
create_folds(
  y,
  k = 5,
  type = c("stratified", "basic", "grouped"),
  m_repetitions = 1,
  names = TRUE,
  return_out = FALSE,
  seed = NULL
)
}
\arguments{
\item{y}{Either the variable used for stratification or grouping.}

\item{k}{Number of folds.}

\item{type}{Split type. One of "stratified", "basic", "grouped".}

\item{m_repetitions}{How many times should the data be split into k folds? Default is 1, i.e. no repeated folds.}

\item{names}{Should folds be named?}

\item{return_out}{Set to \code{TRUE} if the row numbers not in the fold are to be returned. Defaults to \code{FALSE}.}

\item{seed}{Integer random seed.}
}
\value{
A list with row numbers per fold.
}
\description{
This function provides a list of row numbers per fold of k-fold cross-validation (basic, stratified, or grouped). Repeated fold creation is supported as well.
}
\examples{
y <- rep(c(letters[1:4]), each = 5)
create_folds(y)
create_folds(y, k = 2)
create_folds(y, k = 2, m_repetitions = 2)
}
