% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_split.R
\name{make_split}
\alias{make_split}
\title{Split data}
\usage{
make_split(
  y,
  p,
  type = c("stratified", "basic", "grouped"),
  invert = FALSE,
  seed = NULL
)
}
\arguments{
\item{y}{Either the variable used for stratification or grouping.}

\item{p}{A vector with split probabilities, e.g. c(train = 0.7, valid = 0.3).}

\item{type}{Split type. One of "stratified", "basic", "grouped".}

\item{invert}{Set to \code{TRUE} if the row numbers not in the split are to be returned. Defaults to \code{FALSE}.}

\item{seed}{Integer random seed.}
}
\value{
A list with row numbers.
}
\description{
This function provides row numbers per data split, see examples, e.g. into training/validation/test. Different types of splits are supported (basic, stratified, or grouped).
}
\examples{
y <- rep(c(letters[1:4]), each = 5)
make_split(y, p = c(0.7, 0.3))
make_split(y, p = c(train = 0.8, valid = 0.1, test = 0.1))
make_split(rep(c("a", "b"), each = 10), p = c(0.7, 0.3), type = "grouped", seed = 30)
}
