% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_strata.R
\name{multi_strata}
\alias{multi_strata}
\title{Create strata from multiple features}
\usage{
multi_strata(df, num_cat = 3L)
}
\arguments{
\item{df}{A data.frame. Interactions of all provided columns are computed on
which the splitting is then performed (see Details).}

\item{num_cat}{An integer to define the number of groups to categorize
numeric variables when perform splitting on multiple criteria
(Default: \code{NULL}). See Details for further information.}
}
\value{
A vector with \code{m} groups created using interactions between the
features from the provided data.frame that can be passed further on to
the splitting functions.
}
\description{
This function is a helper function to create strata based on multiple criteria,
which can then be used as input to the splitting functions.
}
\details{
The interactions of all columns are calculated to define
\code{m} 'groups', which can be used to perform  the splitting on.
This enables splitting by multiple criteria. While columns of a type
other than \code{numeric} are passed unmodified to \code{interaction},
a vector of quantiles is computed for numeric variables to group /
categorize the values before passing them further on.
The number of categories to split numeric variables is to be provided
with the argument \code{multi_num_cat}.
}
\examples{
y_multi <- data.frame(
  rep(c(letters[1:4]), each = 100),
  factor(sample(c(0, 1), 400, replace = TRUE)),
  rnorm(400)
)
multi_strata(y_multi, num_cat = 3)
}
\seealso{
\code{\link[=interaction]{interaction()}}, \code{\link[=quantile]{quantile()}}, \code{\link[=cut]{cut()}}
}
